Reorganize search components.
Remove modal folder and components.
