+++ app/assets/javascripts/components/layout/SearchBox.jsx
