+++ app/assets/javascripts/components/pages/Search/SearchFacets.jsx
