+++ app/assets/javascripts/components/pages/Search/SearchPagination.jsx
