+++ app/assets/javascripts/components/pages/Search/SearchSidebar.jsx
